+++ lib/grape/middleware/error.rb
@@ -61,7 +61,7 @@ module Grape
       def format_message(message, backtrace)
         format = env['api.format'] || options[:format]
         formatter = Grape::ErrorFormatter::Base.formatter_for(format, options)
-        throw :error, :status => 406, :message => "The requested format #{format} is not supported." unless formatter
+        throw :error, :status => 406, :message => "The requested format '#{format}' is not supported." unless formatter
         formatter.call(message, backtrace, options, env)
       end