+++ lib/grape/middleware/versioner/header.rb
@@ -29,7 +29,7 @@ module Grape
           if strict?
             # If no Accept header:
             if header.qvalues.empty?
-              throw :error, :status => 406, :headers => {'X-Cascade' => 'pass'}, :message => 'Accept header must be set'
+              throw :error, :status => 406, :headers => {'X-Cascade' => 'pass'}, :message => 'Accept header must be set.'
             end
             # Remove any acceptable content types with ranges.
             header.qvalues.reject! do |media_type,_|
@@ -37,7 +37,7 @@ module Grape
             end
             # If all Accept headers included a range:
             if header.qvalues.empty?
-              throw :error, :status => 406, :headers => {'X-Cascade' => 'pass'}, :message => 'Accept header must not contain ranges ("*")'
+              throw :error, :status => 406, :headers => {'X-Cascade' => 'pass'}, :message => 'Accept header must not contain ranges ("*").'
             end
           end
 
@@ -57,8 +57,8 @@ module Grape
           elsif strict?
             throw :error, :status => 406, :headers => {'X-Cascade' => 'pass'}, :message => '406 Not Acceptable'
           # If all acceptable content types specify a vendor or version that doesn't exist:
-          elsif header.values.all?{|media_type| has_vendor?(media_type) || has_version?(media_type)}
-            throw :error, :status => 406, :headers => {'X-Cascade' => 'pass'}, :message => 'API vendor or version not found'
+          elsif header.values.all?{ |media_type| has_vendor?(media_type) || has_version?(media_type)}
+            throw :error, :status => 406, :headers => {'X-Cascade' => 'pass'}, :message => 'API vendor or version not found.'
           end
         end